pyproject.toml
[project]
name = "toolscreen-discord-bot"
version = "0.1.0"
description = "Discord bot for Toolscreen support triage"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.6",
    "pyyaml>=6.0",
]

[[project.authors]]
name = "fe-art"
